requirements_live.txt

fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
redis
PyJWT
openai
webrtcvad-python
deepgram-sdk
numpy
python-multipart
python-jose[cryptography]
websockets
